requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
boto3==1.35.62
    # via -r requirements.in
botocore==1.35.62
    # via
    #   boto3
    #   s3transfer
build==1.2.2.post1
    # via pip-tools
chardet==5.2.0
    # via reportlab
click==8.1.7
    # via pip-tools
cssselect2==0.7.0
    # via svglib
dj-database-url==2.3.0
    # via -r requirements.in
django==5.1.3
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-classy-tags
    #   django-cms
    #   django-entangled
    #   django-filer
    #   django-formtools
    #   django-parler
    #   django-polymorphic
    #   django-sekizai
    #   django-storage-url
    #   django-storages
    #   django-treebeard
    #   djangocms-admin-style
    #   djangocms-alias
    #   djangocms-frontend
    #   djangocms-versioning
    #   easy-thumbnails
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==4.1.4
    # via
    #   -r requirements.in
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-frontend
    #   djangocms-text
    #   djangocms-versioning
django-entangled==0.5.4
    # via djangocms-frontend
django-filer==3.2.3
    # via
    #   -r requirements.in
    #   djangocms-frontend
django-formtools==2.5.1
    # via django-cms
django-fsm==2.8.2
    # via djangocms-versioning
django-parler==2.3
    # via djangocms-alias
django-polymorphic==3.1.0
    # via django-filer
django-sekizai==4.1.0
    # via django-cms
django-storage-url==0.12.1
    # via -r requirements.in
django-storages==1.14.4
    # via django-storage-url
django-treebeard==4.7.1
    # via django-cms
djangocms-admin-style==3.3.1
    # via
    #   -r requirements.in
    #   django-cms
djangocms-alias==2.0.1
    # via -r requirements.in
djangocms-attributes-field==3.0.0
    # via djangocms-frontend
djangocms-frontend==1.3.5
    # via -r requirements.in
djangocms-text==0.3.3
    # via -r requirements.in
djangocms-versioning==2.1.0
    # via -r requirements.in
easy-thumbnails[svg]==2.10
    # via
    #   -r requirements.in
    #   django-filer
    #   djangocms-frontend
furl==2.1.3
    # via django-storage-url
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==5.3.0
    # via
    #   djangocms-text
    #   svglib
nh3==0.2.18
    # via djangocms-text
orderedmultidict==1.0.1
    # via furl
packaging==24.2
    # via
    #   build
    #   django-cms
    #   djangocms-text
pillow==11.0.0
    # via
    #   djangocms-text
    #   easy-thumbnails
    #   reportlab
pip-tools==7.4.1
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via botocore
reportlab==4.2.5
    # via
    #   easy-thumbnails
    #   svglib
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
sqlparse==0.5.2
    # via django
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.4.0
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.12.2
    # via dj-database-url
urllib3==2.2.3
    # via botocore
uwsgi==2.0.28
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
wheel==0.45.0
    # via pip-tools
whitenoise==6.8.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools


# Core
#python-dotenv==1.0.1
str2bool==1.1

# UI
django-theme-material-kit==1.0.25

# Profiles
django-userena-ce==7.0.3

django-debug-toolbar